+++ templates/clicker-game-tg-mini-app-template/frontend/entry-functions/click.ts
@@ -1,0 +1,11 @@
+import { MODULE_ADDRESS } from "@/constants";
+import { InputTransactionData } from "@aptos-labs/wallet-adapter-react";
+
+export const click = (): InputTransactionData => {
+  return {
+    data: {
+      function: `${MODULE_ADDRESS}::counter_app::click`,
+      functionArguments: [],
+    },
+  };
+};