+++ templates/clicker-game-tg-mini-app-template/frontend/main.tsx
@@ -1,0 +1,27 @@
+// Uncomment this import in case, you would like to develop the application even outside
+// the Telegram application, just in your browser.
+// import "./mockEnv.ts";
+
+import "./index.css";
+
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+import App from "@/App.tsx";
+// Internal components
+import { Toaster } from "@/components/ui/toaster.tsx";
+import { WalletProvider } from "@/components/WalletProvider.tsx";
+
+const queryClient = new QueryClient();
+  
+ReactDOM.createRoot(document.getElementById("root")!).render(
+  <React.StrictMode>
+    <WalletProvider>
+      <QueryClientProvider client={queryClient}>
+        <App />
+        <Toaster />
+      </QueryClientProvider>
+    </WalletProvider>
+  </React.StrictMode>,
+);